+++ src/delete.c
@@ -27,7 +27,7 @@ void delete_work(const char *cal_path, DF works[static 1], const char *opt) {
 	}
 
 	// Find Work Row
-	uint_fast32_t sel_row = select_row(works, opt);
+	size_t sel_row = select_row(works, opt);
 	if (sel_row == 0) {
 		puts("no works match that name, cancelling");
 		puts("=========================================================");
@@ -40,7 +40,7 @@ void delete_work(const char *cal_path, DF works[static 1], const char *opt) {
 	char row_buff[MAX_ROW_LEN + 1]; // +1 for '\n'
 	// Loop Through Until sel_row
 	// (no need for -1 bc of column row)
-	uint_fast32_t i;
+	size_t i;
 	for (i = 0; i < sel_row; i++) {
 		fgets(row_buff, MAX_ROW_LEN + 1, works->fp);
 		fwrite(row_buff, 1, strlen(row_buff), temp_fp);
